dcp-712 Move EntityMap test suite to its own file
